// FrontEnd/ChamberAppel/src/app/generatedapis/models/DisciplineBudgetaireTempDatatableResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { DisciplineBudgetaireTemp } from './DisciplineBudgetaireTemp';
export type DisciplineBudgetaireTempDatatableResponse = {
    data?: Array<DisciplineBudgetaireTemp> | null;
    recordTotal?: number;
    recordFiltred?: number;
};